[REF] {purchase_,}stock,{sale_,}mrp: remove common2.py file
