+++ addons/purchase_stock/tests/common.py
@@ -2,7 +2,7 @@
 from datetime import timedelta
 
 from odoo import fields
-from odoo.addons.stock.tests.common2 import TestStockCommon
+from odoo.addons.stock.tests.common import TestStockCommon
 from odoo import tools
