+++ addons/sale_mrp/tests/test_sale_mrp_lead_time.py
@@ -4,7 +4,7 @@
 from datetime import timedelta
 
 from odoo import fields
-from odoo.addons.stock.tests.common2 import TestStockCommon
+from odoo.addons.stock.tests.common import TestStockCommon
 
 from odoo.tests import Form
