+++ addons/stock/tests/common.py
@@ -1,10 +1,22 @@
 # -*- coding: utf-8 -*-
 
 import re
-from odoo.tests import common
+
+from odoo.addons.mail.tests.common import mail_new_test_user
+from odoo.addons.product.tests.common import TestProductCommon
 
 
-class TestStockCommon(common.TransactionCase):
+class TestStockCommon(TestProductCommon):
+    def _create_move(self, product, src_location, dst_location, **values):
+        # TDE FIXME: user as parameter
+        Move = self.env['stock.move'].with_user(self.user_stock_manager)
+        # simulate create + onchange
+        move = Move.new({'product_id': product.id, 'location_id': src_location.id, 'location_dest_id': dst_location.id})
+        move._onchange_product_id()
+        move_values = move._convert_to_write(move._cache)
+        move_values.update(**values)
+        return Move.create(move_values)
+
     @classmethod
     def setUpClass(cls):
         super(TestStockCommon, cls).setUpClass()
@@ -79,7 +91,7 @@ class TestStockCommon(common.TransactionCase):
         cls.uom_unit = cls.env['uom.uom'].search([('category_id', '=', cls.categ_unit), ('uom_type', '=', 'reference')], limit=1)
         cls.uom_unit.write({
             'name': 'Test-Unit',
-            'rounding': 1.0})
+            'rounding': 0.001})
         cls.uom_dozen = cls.UomObj.create({
             'name': 'Test-DozenA',
             'category_id': cls.categ_unit,
@@ -111,6 +123,58 @@ class TestStockCommon(common.TransactionCase):
             (4, cls.env.ref('base.group_multi_company').id),
             (4, cls.env.ref('stock.group_production_lot').id),
         ]})
+        #######################################################################
+        # TODO: refactor these changes from common2.py
+        #######################################################################
+        # User Data: stock user and stock manager
+        cls.user_stock_user = mail_new_test_user(
+            cls.env,
+            name='Pauline Poivraisselle',
+            login='pauline',
+            email='p.p@example.com',
+            notification_type='inbox',
+            groups='stock.group_stock_user',
+        )
+        cls.user_stock_manager = mail_new_test_user(
+            cls.env,
+            name='Julie Tablier',
+            login='julie',
+            email='j.j@example.com',
+            notification_type='inbox',
+            groups='stock.group_stock_manager',
+        )
+
+        # Warehouses
+        cls.warehouse_1 = cls.env['stock.warehouse'].create({
+            'name': 'Base Warehouse',
+            'reception_steps': 'one_step',
+            'delivery_steps': 'ship_only',
+            'code': 'BWH'})
+
+        # Locations
+        cls.location_1 = cls.env['stock.location'].create({
+            'name': 'TestLocation1',
+            'posx': 3,
+            'location_id': cls.warehouse_1.lot_stock_id.id,
+        })
+
+        # Partner
+        cls.partner_1 = cls.env['res.partner'].create({
+            'name': 'Julia Agrolait',
+            'email': 'julia@agrolait.example.com',
+        })
+
+        # Product
+        cls.product_3 = cls.env['product.product'].create({
+            'name': 'Stone',  # product_3
+            'uom_id': cls.uom_dozen.id,
+            'uom_po_id': cls.uom_dozen.id,
+        })
+
+        # Existing data
+        cls.existing_inventories = cls.env['stock.quant'].search([('inventory_quantity', '!=', 0.0)])
+        cls.existing_quants = cls.env['stock.quant'].search([])
+
 
     def url_extract_rec_id_and_model(self, url):
         rec_id = re.findall(r'[?&]id=([^&]+).*', url)
